Chore/reorg payer data import to bundle ops/plrs 651 (#512)
* Correct misspelling in validateQualityOfAccess, change key from allowedPtps to strictlyRequiredPtps

* Update test for validateQualityOfAccess key change to strictlyRequiredPtps

* Tweak util naming for sheet name string check and getAllowedComboHashes

* Change control flow in uploadSheet to pass entire three payer sheets to importPayerHistoricalAccessData after sanitization/validation

* Refactor importPayerHistoricalData to fetch project PTPs, validate all payer sheets, upsert all, then materialize

* Tweak comments

* Update changelog

* Correct changelog for previous misplacement of entry about DB_CLUSTER_ENV

+++ src/backend/resolvers/mutations/collection/uploadSheet/uploadSheet.js
@@ -4,22 +4,31 @@ const getSheetConfig = require('./utils/getSheetConfig')
 const formatAjvErrors = require('./utils/formatAjvErrors')
 const importPayerHistoricalAccessData = require('./importPayerHistoricalAccessData')
 
+const {
+  isQualityAccessSheet,
+  isAdditionalCriteriaSheet,
+  isPolicyLinksSheet,
+} = require('./importPayerHistoricalAccessData/utils')
+
 const uploadSheet = async (
   parent,
   { input }, // schema is [ { wb, sheet, data, timestamp, projectId }, etc. ]
   { pulseCoreDb, pulseDevDb, mongoClient },
   info
 ) => {
-  const importFeedback = []
+  // Step 1: Completely sanitize and validate the data (across all incoming sheets)
+  // If anything fails, error is thrown right away and no further code executes.
+  const cleanedSheetsWithMetadata = []
 
   for (const sheetToUpload of input) {
-    let { wb, sheet, data, timestamp, projectId } = sheetToUpload
+    let { wb, sheet, data } = sheetToUpload
     const originalDataLength = data.length // excludes header
 
     const { result, skippedRows } = sanitize(data)
     data = result
 
     const sheetConfig = await getSheetConfig({ wb, sheet, pulseCoreDb }) // handles getting the right sheet config, including for payer historical data exceptions
+    
     const targetCollection = sheetConfig.collection
 
     const {
@@ -35,37 +44,59 @@ const uploadSheet = async (
       throw new Error(errorString)
     }
 
-    if (timestamp && projectId) {
-      await importPayerHistoricalAccessData(
-        {
-          wb,
-          sheet,
-          data,
-          timestamp,
-          projectId,
-        },
-        { pulseCoreDb, pulseDevDb, mongoClient },
-        importFeedback,
-      )
+    cleanedSheetsWithMetadata.push({
+      ...sheetToUpload, // if this is payer historical access import, there'll be projectId and timestamp in here
+      data,
+      skippedRows,
+      originalDataLength,
+      targetCollection,
+    })
+  }
 
-      continue
-    }
+  // Step 2: Detect whether the incoming data is a payer workbook or
+  // a standalone sheet; if it's payer workbook, for now let's technically allow
+  // someone to be able to import fewer than 3 sheets (that might be wanted).
+  //  - If payer workbook, use the importPayerHistoricalAccessData pipeline;
+  //  - Else persist straight to pulse-dev for now.
+  const importFeedback = []
 
-    // for data going straight to dev, tack on createdOn to every doc
-    const createdOn = new Date()
-    data = data.map(datum => ({ ...datum, createdOn }))
+  const isIncomingDataPayerWorkbook = isQualityAccessSheet(input[0].sheet)
+    || isAdditionalCriteriaSheet(input[0].sheet)
+    || isPolicyLinksSheet(input[0].sheet)
 
-    await pulseDevDb.collection(targetCollection)
-      .deleteMany()
+  if (isIncomingDataPayerWorkbook) {
+    await importPayerHistoricalAccessData({
+      cleanedSheetsWithMetadata,
+      dbsConfig: { pulseCoreDb, pulseDevDb, mongoClient },
+      importFeedback,
+    })
+  } else {
+    for (const sheetObjWithMetadata of cleanedSheetsWithMetadata) {
+      let {
+        wb,
+        sheet,
+        data,
+        skippedRows,
+        originalDataLength,
+        targetCollection,
+      } = sheetObjWithMetadata
 
-    await pulseDevDb.collection(targetCollection)
-      .insertMany(data)
+      // for data going straight to dev, tack on createdOn to every doc
+      const createdOn = new Date()
+      data = data.map(datum => ({ ...datum, createdOn }))
 
-    importFeedback.push(
-      `Import successful for ${wb}/${sheet}`
+      await pulseDevDb.collection(targetCollection)
+        .deleteMany()
+
+      await pulseDevDb.collection(targetCollection)
+        .insertMany(data)
+
+      importFeedback.push(
+        `Import successful for ${wb}/${sheet}`
         + `\n${data.length}/${originalDataLength} rows imported (excluding header)`
         + `\nSkipped rows were: ${skippedRows.join(', ')}`
-    )
+      )
+    }
   }
 
   return importFeedback
